+++ src/surrealist/ql/statements/define.py
@@ -1,10 +1,11 @@
+import json
 from abc import ABC
 from typing import List, Union, Any, Optional, Tuple
 
 from surrealist.connections import Connection
 from surrealist.ql.statements.define_index_statements import CanUseIndexTypes
 from surrealist.ql.statements.permissions import CanUsePermissions
-from surrealist.ql.statements.statement import Statement
+from surrealist.ql.statements.statement import Statement, FinishedStatement
 from surrealist.utils import OK
 
 
@@ -16,9 +17,13 @@ class Define(Statement, ABC):
     def __init__(self, connection: Connection):
         super().__init__(connection)
         self._if_not_exists = False
+        self._text = None
 
     def _exists(self) -> str:
         return " IF NOT EXISTS" if self._if_not_exists else ""
+
+    def _comment(self) -> str:
+        return f" COMMENT {json.dumps(self._text)}" if self._text else ""
 
     def if_not_exists(self) -> "Define":
         """
@@ -28,6 +33,15 @@ class Define(Statement, ABC):
         self._if_not_exists = True
         return self
 
+    def comment(self, comment: str) -> FinishedStatement:
+        """
+        Adds COMMENT statement to the query
+        :param comment: comment text
+        :return: self
+        """
+        self._text = comment
+        return FinishedStatement(self)
+
 
 class DefineEvent(Define):
     """
@@ -36,6 +50,8 @@ class DefineEvent(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/event
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE EVENT [ IF NOT EXISTS ] @name ON [ TABLE ] @table [ WHEN @expression ] THEN @expression [ COMMENT @string ]
 
     """
 
@@ -64,7 +80,8 @@ class DefineEvent(Define):
 
     def _clean_str(self):
         when = f"WHEN {self._when} " if self._when else ""
-        return f"DEFINE EVENT{self._exists()} {self._name} ON TABLE {self._table_name} {when}THEN {self._then}"
+        return f"DEFINE EVENT{self._exists()} {self._name} ON TABLE {self._table_name} {when}" \
+               f"THEN {self._then}{self._comment()}"
 
 
 class DefineUser(Define):
@@ -74,6 +91,9 @@ class DefineUser(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/user
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE USER [ IF NOT EXISTS ] @name ON [ ROOT | NAMESPACE | DATABASE ] [ PASSWORD @pass | PASSHASH @hash ]
+    ROLES @roles [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, user_name: str, password: str):
@@ -106,7 +126,8 @@ class DefineUser(Define):
         return [OK]
 
     def _clean_str(self):
-        return f"DEFINE USER{self._exists()} {self._name} ON DATABASE PASSWORD '{self._pass}' ROLES {self._role}"
+        return f"DEFINE USER{self._exists()} {self._name} ON DATABASE PASSWORD '{self._pass}' " \
+               f"ROLES {self._role}{self._comment()}"
 
 
 class DefineParam(Define):
@@ -116,6 +137,8 @@ class DefineParam(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/param
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE PARAM [ IF NOT EXISTS ] $@name VALUE @value [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, param_name: str, value: Any):
@@ -131,7 +154,7 @@ class DefineParam(Define):
         return [OK]
 
     def _clean_str(self):
-        return f"DEFINE PARAM{self._exists()} ${self._name} VALUE {self._value}"
+        return f"DEFINE PARAM{self._exists()} ${self._name} VALUE {self._value}{self._comment()}"
 
 
 class DefineAnalyzer(Define):
@@ -141,6 +164,8 @@ class DefineAnalyzer(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/analyzer
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE ANALYZER [ IF NOT EXISTS ] @name [ TOKENIZERS @tokenizers ] [ FILTERS @filters ] [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, name: str):
@@ -173,7 +198,7 @@ class DefineAnalyzer(Define):
     def _clean_str(self):
         tok = "" if not self._tokenizers else f" TOKENIZERS {self._tokenizers}"
         filters = "" if not self._filters else f" FILTERS {self._filters}"
-        return f"DEFINE ANALYZER{self._exists()} {self._name}{tok}{filters}"
+        return f"DEFINE ANALYZER{self._exists()} {self._name}{tok}{filters}{self._comment()}"
 
 
 class DefineScope(Define):
@@ -183,6 +208,8 @@ class DefineScope(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/scope
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE SCOPE [ IF NOT EXISTS ] @name SESSION @duration SIGNUP @expression SIGNIN @expression [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, name: str, duration: str, signup: Union[str, Statement],
@@ -204,7 +231,7 @@ class DefineScope(Define):
 
     def _clean_str(self):
         return f"DEFINE SCOPE{self._exists()} {self._name} SESSION {self._duration} \nSIGNUP {self._signup} " \
-               f"\nSIGNIN {self._signin}"
+               f"\nSIGNIN {self._signin}{self._comment()}"
 
 
 class DefineIndex(Define, CanUseIndexTypes):
@@ -214,6 +241,14 @@ class DefineIndex(Define, CanUseIndexTypes):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/indexes
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE INDEX [ IF NOT EXISTS ] @name ON [ TABLE ] @table [ FIELDS | COLUMNS ] @fields
+    [ UNIQUE
+        | SEARCH ANALYZER @analyzer [ BM25 [(@k1, @b)] ] [ HIGHLIGHTS ]
+        | MTREE DIMENSION @dimension [ TYPE @type ] [ DIST @distance ] [ CAPACITY @capacity]
+        | HNSW DIMENSION @dimension [ TYPE @type ] [DIST @distance] [ EFC @efc ] [ M @m ]
+    ]
+    [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, name: str, table_name: str):
@@ -252,7 +287,7 @@ class DefineIndex(Define, CanUseIndexTypes):
         return [OK]
 
     def _clean_str(self):
-        return f"DEFINE INDEX{self._exists()} {self._name} ON TABLE {self._table_name} {self._fields}"
+        return f"DEFINE INDEX{self._exists()} {self._name} ON TABLE {self._table_name} {self._fields}{self._comment()}"
 
 
 class DefineToken(Define):
@@ -262,6 +297,9 @@ class DefineToken(Define):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/token
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE TOKEN [ IF NOT EXISTS ] @name ON [ NAMESPACE | DATABASE | SCOPE @scope ] TYPE @type VALUE @value
+    [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, name: str, token_type: str, value: str):
@@ -278,7 +316,8 @@ class DefineToken(Define):
         return [OK]
 
     def _clean_str(self):
-        return f'DEFINE TOKEN{self._exists()} {self._name} ON DATABASE \nTYPE {self._type} \nVALUE "{self._value}"'
+        return f'DEFINE TOKEN{self._exists()} {self._name} ON DATABASE \nTYPE {self._type} \n' \
+               f'VALUE "{self._value}"{self._comment()}'
 
 
 class DefineTable(Define, CanUsePermissions):
@@ -288,6 +327,24 @@ class DefineTable(Define, CanUsePermissions):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/table
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+    
+    DEFINE TABLE [ IF NOT EXISTS ] @name
+    [ DROP ]
+    [ SCHEMAFULL | SCHEMALESS ]
+    [ TYPE [ ANY | NORMAL | RELATION [ IN | FROM ] @table [ OUT | TO ] @table ] ]
+    [ AS SELECT @projections
+        FROM @tables
+        [ WHERE @condition ]
+        [ GROUP [ BY ] @groups ]
+    ]
+    [CHANGEFEED @duration [INCLUDE ORIGINAL] ]
+    [ PERMISSIONS [ NONE | FULL
+        | FOR select @expression
+        | FOR create @expression
+        | FOR update @expression
+        | FOR delete @expression
+    ] ]
+    [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, name: str):
@@ -411,7 +468,7 @@ class DefineTable(Define, CanUsePermissions):
             if self._changefeed[1]:
                 feed = f"{feed} INCLUDE ORIGINAL"
         type_ = "" if not self._type else f" TYPE {self._type}"
-        return f'DEFINE TABLE{self._exists()} {self._name}{drop}{schema}{type_}{alias}{feed}'
+        return f'DEFINE TABLE{self._exists()} {self._name}{drop}{schema}{type_}{alias}{feed}{self._comment()}'
 
 
 class DefineField(Define, CanUsePermissions):
@@ -421,6 +478,20 @@ class DefineField(Define, CanUsePermissions):
     Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/field
 
     Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
+
+    DEFINE FIELD [ IF NOT EXISTS ] @name ON [ TABLE ] @table
+    [ [ FLEXIBLE ] TYPE @type ]
+    [ DEFAULT @expression ]
+  [ READONLY ]
+    [ VALUE @expression ]
+    [ ASSERT @expression ]
+    [ PERMISSIONS [ NONE | FULL
+        | FOR select @expression
+        | FOR create @expression
+        | FOR update @expression
+        | FOR delete @expression
+    ] ]
+  [ COMMENT @string ]
     """
 
     def __init__(self, connection: Connection, field_name: str, table_name: str):
@@ -486,4 +557,4 @@ class DefineField(Define, CanUsePermissions):
         value = "" if not self._value else f" VALUE {self._value}"
         asserts = "" if not self._assert else f" ASSERT {self._assert}"
         return f'DEFINE FIELD{self._exists()} {self._field_name} ' \
-               f'ON TABLE {self._table_name}{type_}{default}{ro}{value}{asserts}'
+               f'ON TABLE {self._table_name}{type_}{default}{ro}{value}{asserts}{self._comment()}'